progress on heat map

+++ WeatherKitDemo/Extensions/WeatherServiceExtension.swift
@@ -3,7 +3,7 @@ import SwiftUI
 import WeatherKit
 
 extension WeatherService {
-    static let days = 5.0
+    static let days = 5
 
     func hourlyForecast(for location: CLLocation) async -> [HourWeather] {
         let forecast = try? await weather(
@@ -28,7 +28,9 @@ extension WeatherService {
 
         // Only keep forecasts in the future, not more a given number of days.
         let startDate = Date()
-        let endDate = startDate.addingTimeInterval(Self.days * 24 * 60 * 60)
+        let endDate = startDate.addingTimeInterval(
+            Double(Self.days * 24 * 60 * 60)
+        )
 
         let futureForecast = forecast.filter {
             startDate <= $0.date && $0.date <= endDate
